Enable recursion_depth=2 + fix fp16 softmax dtype in sparse_attn
- ModelConfig700M: recursion_depth 1→2 (now safe with fp16 model ~3.3GB)
  TRM-style: full layer stack applied twice per forward pass, doubling
  effective compute depth with zero extra parameters.

- sparse_attn.py: softmax upcast to fp32 then cast back to V.dtype.
  F.softmax always returns float32; multiplying float32 attn × fp16 V
  raised RuntimeError outside of autocast context.

+++ arch/sparse_attn.py
@@ -205,7 +205,7 @@ class SparseAttention(nn.Module):
         scale = math.sqrt(d_head)
         attn = torch.matmul(Q, K.transpose(-2, -1)) / scale    # (B, H, L, total_kv)
         attn = attn + mask.unsqueeze(0).unsqueeze(0)
-        attn = F.softmax(attn, dim=-1)
+        attn = F.softmax(attn.float(), dim=-1).to(V.dtype)     # fp32 softmax, cast back
 
         out = torch.matmul(attn, V)                             # (B, H, L, dh)
         out = out.permute(0, 2, 1, 3).reshape(B, L, n_heads * d_head)
